+++ app/dta-sender.js
@@ -13,22 +13,55 @@ exports: module.exports = {
   sendDta,
   sendLogs,
   sendAccessStats,
-  getSendErrCnt
+  getStats
 }
 
+let sndStats = {
+  sentDtaCnt: 0,
+  sentLogsCnt: 0,
+  sentAStsCnt: 0,
+  errCnt: 0
+}
 async function sendDta( dta ) {
+  sndStats.sentDtaCnt ++
   return await send( dta, '/mon/dta' )
 }
 
+let resendLogs = []
 async function sendLogs( dta ) {
-  return await send( dta, '/mon/logs' )
+  while ( resendLogs.length > 0 ) {
+    let snd = await send( resendLogs[0], '/mon/logs' )
+    if ( snd.error ) { return }
+    log.info( '... resend logs: OK' )
+    resendLogs.shift()
+  }
+  let snd = await send( dta, '/mon/logs' )
+  if ( snd.error ) {
+    resendLogs.push( dta )
+  } else {
+    sndStats.sentLogsCnt ++
+  }
+  return snd
 }
 
+
+let resendStats = []
 async function sendAccessStats( dta ) {
-  return await send( dta, '/mon/access-stats' )
+  while ( resendStats.length > 0 ) {
+    let snd = await send( resendStats[0], '/mon/access-stats' )
+    if ( snd.error ) { return }
+    log.info( '... resend ingress stats: OK' )
+    resendStats.shift()
+  }
+  let snd = await send( dta, '/mon/access-stats' )
+  if ( snd.error ) {
+    resendStats.push( dta )
+  } else {
+    sndStats.sentAStsCnt ++
+  }
+  return snd
 }
 
-let errCnt = 0
 
 async function send( dta, path ) {
   return new Promise( ( resolve, reject ) => {
@@ -45,26 +78,25 @@ async function send( dta, path ) {
       ).then( req => {
         if ( req.request.res.statusCode != 200 ) {
           log.warn( sendDta, req.request.res.statusMessage )
-          errCnt ++
+          sndStats.errCnt ++
         }
         log.verbose( 'send res', req.data  )
         resolve( req.data )
       }).catch( error => {
         log.warn( 'send', path, error.message )
-        errCnt ++
+        sndStats.errCnt ++
         resolve({ error:error.message })
       })
     } catch ( exc ) {
       log.error( 'send', path, exc.message )
-      errCnt ++
+      sndStats.errCnt ++
       resolve({ error:error.message })
     }
   })
 }
 
-function getSendErrCnt() {
-  let result =0
-  result += errCnt
-  errCnt = 0
+function getStats() {
+  let result = JSON.parse( JSON.stringify( sndStats ) )
+  for ( let key in sndStats ) { sndStats[ key ] = 0 }
   return result
 }